+++ res/colors.js
@@ -3,7 +3,14 @@ const GREEN   = '#4CAF50'; //positive // IF CHANGED : Check getResults.js
 const YELLOW  = '#FFEB3B'; //warning
 const BLUE    = '#919AB1'; //neutral
 const BLACK   = '#000000'; //muted
-const COUNCIL = '#9c3848'; //rick astley
+const COUNCIL = '#9c3848'; //crimson
+
+const C32       = '#76C945';
+const C64       = '#76C945';
+const CADDONS   = '#6D6163';
+const CTWEAKS   = '#658430';
+const CMODS     = '#349687';
+const CDUNGEONS = '#DF7C2B';
 
 module.exports = {
 	RED,
@@ -11,5 +18,11 @@ module.exports = {
 	YELLOW,
 	BLUE,
 	BLACK,
-	COUNCIL
+	COUNCIL,
+	C32,
+	C64,
+	CADDONS,
+	CTWEAKS,
+	CMODS,
+	CDUNGEONS
 }